Update loot api integration to loot api python v3.0.0: RRR
What changed was the way loot_api is initialised and the
get_plugin_tags() and get_plugin_cleanliness() functions, which now need
to be passed True as a second argument to retain their previous
function.

Also updated the install_loot_api and mktaglist scripts. They both work
for me now out of the scripts folder. A bunch of file paths assumed the
scripts to be in the root directory (not in scripts). No idea when they
were last working.

Under #   388.

Signed-off-by: MrD <the.ubik@gmail.com>