Fix for refreshBashTags errors after loot api update:
The error itself should be benign. The `eval_lists()` command was
removed in the newer loot api versions. The loot api integration update
did make use of the new way of evaluating conditions (passing `True` as
a second argument to `get_plugin_tags()` and `get_plugin_cleanliness()`
but failed to removed the now unnecessary and failing `eval_lists()`
calls.

Apparently the error would only show for Skyrim/SSE, not for FO4 or
Oblivion.

Under #388.